Scrub #coordinates from response
Now we are using the MediaWiki coordinates
prop to source coordinates there is no need to keep this element in
the HTML. On many pages it may be the first paragraph so removing it
will help us serve the first paragraph of content to our users

After this change:
http://0.0.0.0:6927/en.wikipedia.org/v1/page/preview-html/Spain
will render a sensible summary.

Additional changes:
Selectors are joined into a single querySelectorAll call.
Some selectors like this one may be ids and this approach avoids
the issue with https://github.com/fgnass/domino/issues/99 with making
a document querySelectorAll call on an id (which is just as easily done
by a querySelector call.

Bug: T172021
Change-Id: Ia854f7f6808a745634d9b72e12ea2880fe579c66