Add read-html route
Now using parsoid-dom-utils to add <section> elements and split up the
DOM based on <section> elements. Note that <section> elements can be
nested but we stop compiling section text for the current section when
we encounter the start of next sub-section.

Added check for sectioning in Parsoid. We don't want to repeat sectioning
once Parsoid provides that for us (T114072).

One exception is the section parsing for the Wiktionary definition
endpoint: this one currently uses the old methods (just renamed to
have the 'old' suffix) since adding the <section> elements requires
changing the definition parsing significantly. That's probably worthy
of a separate commit.

Renamed parseSection.js to parsoidSectionsForDivs.js. Also moved over the
other code responsible for dealing with sections.

Bug: T162179
Change-Id: I4d408349f40f9289e63a6f6f3b955cf403f48fc3