Add compare script for old and new extracts
Compare the top 1000 pages viewed on enwiki, and compares
HTML version of the extract of old and new summary implementations.

Arguments: provide a single argument which is the language code for
the Wikipedia project.

MAke sure your local MCS instance is running, then execute the
comparison script with a command like this:
cd scripts
node ./compare-extracts.js en

The output will be three files in the ../private/extracts folder:
* en.html
* en.v1.txt
* en.v2.txt

The HTML file is good for an overview of the visible quality. The two
txt files are better when using a good diff tool, and allow more control
and show details better. Compare the txt files if you are concerned
about how the HTML tags compare.

Bug: T175286
Change-Id: Ia217c9613c8458352882862b43a6005cb6092158