Hygiene: move some files from static to private folder
The static folder is for serving static content.
The files moved in this patch should not be served at all.

Also start a subfolder for the lists of top pages since there
will be many more in the future.

Updated main-page-names.py to use new folder for output.
Bonus: removed an extra semicolon there (doesn't belong in
a Python script anyway).

Change-Id: Iaa438fd19555635c577d3a2bcc40303e641be517