Hygiene: update diff tests
userinfo is removed now. MW API returns en error in this case, because
it our code assumes the user name is the title, even for subpages in
User namespace. In a follow-up patch we should decide how we want to
handle these cases. This patch is just to get the build fixed.

Change-Id: Ia76d2c4c84d1e5e8c5ae53a87085e9aa996f334e