Hygiene: Update test fixtures
This is for news, onthisday, and a general test page.
This ia a followup of the comment in
https://gerrit.wikimedia.org/r/#/c/382035/1//COMMIT_MSG@9.

The updated content is based on an undeployed Parsoid version:
https://gerrit.wikimedia.org/r/#/c/370227/12
to make sure MCS is ready for this.

Using
    const uri = `http://localhost:8000/${lang}.wikipedia.org/v3/page/pagebundle/${encodeURIComponent(NEWS_SITES[lang].title)}`;
Note: pagebundle instead of html in the path.
And used rsp.body.html.body instead of rsp.body when writing.

Change-Id: Ib50b702a37acf6cb481d5ffa731e48e6e8490303