Fix: Get the default set of pageprops in getMetadataActionApi
The MW API parameter for specifying a subset of pageprops to return is
'ppprop', not 'pageprops'.  We have been receiving an API warning and
the default subset.  Actually, this seems to suit our purposes fine.  Less
code is better, so let's just go ahead and keep using the default subset.

Change-Id: I9e390d16a2b60bbcde6acb26f2fb2432fd62f320