Summary 2.0: Add mobile content URLs
We want to avoid making clients redirect to the pages they need.  This
adds content URLs specially formatted for mobile clients in order to
accomplish that.

NOTE: The mobile URL bits are commented out of the spec for now so the
CI tests continue to pass, but should be uncommented once the change
adding the base mobile URL to siteinfo is deployed (this week).

Bug: T170692
Change-Id: I3c4131596f2c22c998fe5802752ae93478b94c03