Hygiene: preserve response objects throughout special handling
In mobile-sections we sometimes rebuilt the response objects while
handling some special cases (Main/File/User page). This leaked the
details of the response objects to these many places and we needed
to take care to recreate all desired fields in the response objects.

With this patch this will change. Instead of recreating the object
it will modify or add new properties in those mentioned special cases
but not try to recreate the whole response object.
This leads to simpler code, and fewer changes when new properties are
introduced.

Change-Id: I23fd30fecd05e23f7ef9a1de2389e6993af504f6