Media: incorporate initial data from page HTML
Adds logic to incorporate data from the page HTML when generating a
media item list via querySelectorAll.  Adds 'type', 'start_time',
and 'end_time' via this path.

As a bonus, since an initial array from querySelectorAll is maintained
throughout processing, and results of the follow-up API call for metadata
are simply merged in, there's no need for a follow-up sort.

Bug: T177430
Change-Id: Ib183edde7c1ecfabe2f37c4289c0471616e2572c