Add .nsprc and update dependencies to resolve build issues
Updated the dependencies that are unique to mobileapps (i.e., not
inherited from the template).

Notably, eslint, eslint-mocha, and ajv needed updating in order to meet
the new peer dependency requirements as a result of a recent change to
eslint-config-node-services.

Fixed indentation problems uncovered by the config update.

Adds an exception to nsp checking for moment.js to unblock us, but we
should wait for a real fix before deploying again.

Bug: T181564
Change-Id: I3110d00dcb9a15914861b5634b504df37df22a33