Stop suppressing RESTBase redirect handling when requesting HTML
Wiki pages occasionally move, and there's no reason not to let RESTBase
handle redirects for us when requesting page HTML. This will make the
service more resilient and improve the stability of our tests.

Updates getMetadataActionApi to return whether the page is a redirect,
so that the summary endpoint can continue to return 204s for redirect
pages per the spec. (Previously, we were relying on the redirect response
not having any results when querying for selector 'body > p'.)

Change-Id: Id42b6aa3c9fe751de74e6cafedfd622811b0c232