Summary: Send empty extracts, not 204, for main pages
I believe this was added based on a misreading of the earlier patch for
handling redirects.  Instead, main pages should be handled in the general
case of pages with no lead intros, which is to return a valid summary
object with empty extract fields.

This is subject to confirmation from Reading Web but I'm putting in the
patch now since it's nearly deployment time.

Bug: T177431
Change-Id: I3b5448369bd07580842188305d2c246b9f78b111