References: citation type
Instead of an citations array we return a single value in the type
field.
We still collect all citation types in a set.
If there is exactly one type left and it's one of the known values
("web", "news", "journal", or "book") then that value is returned.
Else "generic" is returned.

Bug: T182652
Change-Id: Ic703fbdf642661481b53e3af21eba1a3583366a0