Include H1 tags when evaluating sections
After excluding sections with -1 and -2 IDs from warnings, all of the
remaining warnings I've seen involve sections with H1 headers, which we've
been excluding.  These are rare in the main namespace (with at least one
exception [1]) but rather common in other namespaces.[2]

This expands the logic to include H1 as well.

[1] https://ru.wikipedia.org/api/rest_v1/page/html/Kawasaki_Ki-61_Hien

[2]
https://en.wikipedia.org/api/rest_v1/page/html/Draft:West_Virginia_High_School_Football_State_Championships
https://en.wikipedia.org/api/rest_v1/page/html/User:Sarahj2107/Dashboard
https://en.wikipedia.org/api/rest_v1/page/html/User:Vicenarian/State_of_the_Wiki

Bug: T182774
Change-Id: I871225870eec478f4e6281b35b9d432a7f63b63e