Show MW API requests when debug is true
To make it easier to debug backend requests I think it is useful
to see the full query string made to the MW API.
There can be tons of query parameters, and it's quite tedious to
compile them manually.
To turn this on just set the DEBUG constant in api-util.js to true
and look for "outgoing request" messages in the log file.

Change-Id: Ibd8196b98d90e0431650b4fabc56ef006af16d10