Fix misleading testing utilities and update tests
The deepEqual and isDeepEqual wrapper methods in test/util/assert.js
confusingly use RegExp matching for string inputs in a way that is not
at all equivalent to === and makes it very easy to write tests resulting
in false positives (see
https://github.com/wikimedia/service-template-node/issues/94, for
example).

This eliminates the alternate string behavior and updates the tests that
are broken by that change.

This should be upstreamed in some form.

Change-Id: Ia9b01b26ac004294e10901736ed3d8e8653be984