Refine description selection
Simplifies the description picking logic away from its current two-
step approach to a single step.

Additionally, drops the secondary fallback behavior of returning the
first available description in the event a description is not
available in either the preferred language or the configured fallback
(English), since users are not likely to have any use for a
description in a totally random language.

Change-Id: Idc86f90674ef1bee4f6bbbb64b8058ed105d97f4