Media: Filter images based on size on the page
Reject if on-page width or height < 64px. Previously, we were using
size (among other criteria) as a heuristic for excluding icons and other
images unrelated to the article content, but basing the exclusion on the
file's original size as reported by the MediaWiki API. The on-page size
is what we actually care about.

Bug: T177430
Change-Id: I9e6521b2b0ab579b97ef852e91c7517be6ff15a5