Dev: Make sure we tag the correct commit in the src repo
Get the corresponding commit in the src submodule
Gets the corresponding commit from the src repo via the submodule
and use that for tagging.
This will avoid tagging a commit that has been merged since deployment.
Yes, unfortunately, sometimes I've done the tagging a bit later than I
should have.

Now that we use the correct commit id this script doesn't need to check
anymore that the commit was made in past 2 hours since it's a weaker
check.

Change-Id: If5879fae9ac874de6f13e89fbc7530de1963abe3