Dev: compare extracts: allow comparing with another local MCS instance
Now that we're getting close to exposing the MCS summary endpoint it is
beneficial to be able to compare the outputs of two different local MCS
instances. One could run a different MCS commit than the other one.

This commit also specifies the revision and tid from the top-pages
files for local URIs. It also adds them to the output files.
For convenience, a link to the new summary URI is also added to the HTML
report.
Also for convenience the CSS file has a line commented out for hiding
same value cells.

Change-Id: I15c78496ef37cf94751a99fe509e919fb2e89613
(cherry picked from commit 200ec32c957376f289d680263e05592a458c7f3e)