Dev: compare extracts: add toggle to show same values
Now that it's likely that the majority of the extract values are the
same it's best to be able to hide the rows with the same value.

Adding a toggle check box to show or hide same value rows.
Consolidated the "same" class to the table row element for more
flexibility.

As a bonus, also show the port numbers in the table heading for old and
new columns.

Change-Id: Icf4978bc5649cfc5fa0108df12a85e5d901b0e36