References: Fix references-large test + parsing of refs in refs
Also specify revisions for better repeatability.
Updated the lookup of the section heading. Needs some more work in
future patches since in some cases we don't get the correct location in
the DOM tree.
Added a single test case to run against a single page which gets use
when the LARGE_TEST environment is not set. The main motivation for this
is to make sure that the general case for this test still works.
Made sure we return promises instead of just creating them to avoid
evergreen tests.

Also fixed an issue where some references inside other references were
not parsed correctly: The ref with id cite_ref-toirupee_6-0 in
https://en.wikipedia.org/wiki/List_of_highest-grossing_Indian_films
did not have the correct backlinks and content.

Continuation of I8147575a19edd62611f844144f062d048e8a08f8.

Change-Id: I991dad035edebdb5daa9e6a9982494c237b81249