Temporarily rewrite sup.mw-ref back to span.mw-ref on mobile-sections*
The Wikipedia Android app currently cannot display inline reference links
in Parsoid HTML, due to having bundled an older version of the
ext.cite.style module that hides the links but can no longer display them
due to a recent Parsoid change.[1]

This patch effectively reverts the Parsoid change for the mobile-sections*
endpoints so that the app shows the inline reference links again.

Note that this will require the app to continue bundling ext.cite.style
for as long as the service-side rewrite is needed.

[1] I7c83d12dd900682799c124ddae1a8689969d5e8c

Bug: T185427
Change-Id: I9d531c14180f9099934bff1997ba6d8cbf18458a