Summary: escape parentheses in attribute values
To avoid parentheses inside attribute values affecting the
parenthetical stripping, which is done on the HTML string via regular
expressions, we want to escape them prior to that, so they don't get
picked up the the regular expressions.

The escapes are done not to the expected representations, for example
we would expect `(` to become `&#x28;`. The reason why that is not done
is because the resulting string we get from innerHTML escapes the `&`,
to we get `&amp;#x28;` instead. Any `&` gets escaped after
innerHTML.

What we do instead is to escape the various parentheses to four
character selected from a reserved Unicode Private Use Area. This
requires that we need to run through this again and unescape those
characters to their original representation.

Bug: T184557
Change-Id: I72ae29dcf1112ff38dde7198344042aef8610ec9