Throw an error if the MediaWiki API responds with text/html
If we get a response from the MediaWiki API with a 200 response code but a
content type of text/html, it's almost certainly the auto-generated API
help page, which means there was a problem with the request.

Currently, this problem is visible if a request is made to mediawiki.org
(without the www subdomain).

Bug: T186745
Change-Id: I0c0630d538b6e62312647b33c294d91341634cba