Dev: Normalize log messages for easier filtering in Kibana
Changed one log message in our code to drop the pid from the message
string since it is already included in the `pid` field.

Changed the string also to start with lower case to be consistent with
the change in https://github.com/wikimedia/service-runner/pull/180.
FWIW, updated the version in package.json. That is mostly for
documentation purposes since we would get that update automatically
anyways.

Change-Id: I1406cfb479f0d689ed4b56b467dc5328a70d68a5