Dev: improve documentation for compare extracts script
Make it easier to start the compare script against production again
by adding more comments in the script for running against 1 or 2 local
MCS instances.

Also stop breaking up words in the CSS by default. I left it in
commented out since it is useful for the comparison of JSON fields.

Change-Id: I12d42ec6cd653c04d98d38fcc3dd036e0615e4cf