Skip transcluded leading non-content paragraphs when extracting lead intro
Fixes an issue seen especially on dewiki in which extractLeadIntroduction
was returning worthless summaries based on text not intended for display.

To do this, we'll leverage the fact that Parsoid includes an "about"
attribute on transcluded content.  Accordingly, we'll need to stop
stripping it from <p> elements.

Updates the earlier workaround for this issue[1] and adds tests.

[1] Ia1343591fe933571815a329e7810261a514e5608

Bug: T186838
Bug: T186847
Change-Id: I5c29668f369573671666453152fe4c5ab26d9167