News: remove floated content
There was an extra square at the beginning of the story text in the
German In the News section today, see task description.
Just removing the <figure-inline> element is not enough.
The square would still show up. So, this just removes any spans with
style float:right. We probably don't need to worry about float:left
since that is not likely to be used in the In the News templates.

Bug: T187475
Change-Id: I883bcb5137a9abbb4d31199d350d28dec6c38c0c