Handle hatnotes and issues in any section
Hatnotes and issues can be present at the top of any section as well as the
top of the page.[1]  This updates extractHatnotes and extractPageIssues to
accommodate this fact, and report hatnotes and issues from all sections in
the metadata endpoint.

New functions extractHatnotesForMobileSections and extractPageIssuesFor
MobileSections are added.  Their output is unchanged from the original
extractHatnotes and extractPageIssues.  As before, they apply only to the
lead section.

For the metadata endpoint, new hatnote and page issue extraction functions
are added to provide the element's textContent in the 'text' property,
innerHTML in the 'html' property, and 'section' in the section property:

  {
    section: 0,
    html: <p>Foo</p>
    text: Foo
  }

The new metadata functions apply to the entire document.

[1] https://en.wikipedia.org/wiki/Wikipedia:Hatnote

Bug: T172002
Bug: T172004
Change-Id: I6b6a6f2cf987f90bb4a30b2384c22cbdf24b9e7b