Summary: keep more (chemical) formulas intact
The Czech version of Nephrite seems fine but the English version
has the problem that there is a space inside the `(Mg, Fe)`.
I think we should keep every parenthetical if there is no space right
before the opening `(`.
This is not done for non-latin variants since the examples we looked at
don't have any leading space before the `（`. (Maybe that's why the
character looks like it has a space before it?)

This was mainly triggered by the desire to keep chemical formulas
but would also help reduce messing with some textual formulas, like
the `10<sup>(10<sup>100</sup>)</sup>` example.

In addition to this we might want to consider adding exceptions of
certain punctuations before spaces not triggering parens removal but
that needs a better approach or it would lead to more complex regular
expressions.

Bug: T183833
Change-Id: Idbae8d0743ca5f35dc37ea243eaa0819ed6a328a