Summary: improve stripping of IPAs
Between the release of 1.3.1 and now we had a regression in terms of
stripping IPAs for the en:London page summary and more. This patch
tries to correct that before we deploy these changes.

Changed the old selector slightly to also trigger when the template
starts with `IPA`, not just equals it.

Change-Id: I215561196ee2b27ebfb8cacde6bf8d0011d4cb98