Summary: flatten spans + fix flattenElement
* Add flattening of spans.
* Some nested spans were not flattened before because of the order the
  replacement of elements was done. Changing the order to from last to
  first fixes this.
* No need for regex to flatten spans with space or No-Break Space
  anymore, but need to drop the extra spaces around them.

Improves stripping of IPAs e.g. for Kyoto article where multiple IPA
templates are used in the same element (and the data-mw for that is on
the parent element.

Between the release of 1.3.1 and now we had a regression in terms of
stripping IPAs for the en:Kyoto page summary and more. This patch
tries to correct that before we deploy these changes.

Change-Id: Ic63023142dcbe738ddabfc0fc1e7cde09e06d49e