Summary: flatten new and redirect links
Removes the classes `new` and `mw-redirect` from <a> elements
before they get turned into `<span>`s. In most cases this results in
a text node instead of the span.

Since the link elements in summaries are changed to span elements the
new and mw-redirect classes don't provide any useful information.
By flattening them we get the html and plain text versions of the
extracts more similar, which improves compression.

Bug: T188005
Change-Id: Id2c4435d80f144fecb8b38d1436226fb976ebc80