Add utility method for evaluating MW API booleans
Handles boolean values regardless of formatversion 1 or 2. This is useful
primarily in situations in which an API modules is returning classic MW
API booleans (i.e., empty strings) despite formatversion 2 being
requested.  In this situation it helps guard against our code suddenly
breaking if the module in question is updated to respect formatversion=2.

Change-Id: I713ac3f507d047449699de92d0a449063a59dc4d