References: add diff test
... so we can easily see how the output changes over time.
The first two also make it easier to keep the documentation page
https://www.mediawiki.org/wiki/Page_Content_Service/References up to
date. I also wanted to add the third example here but the new git hook
doesn't like a special character in the JSON of that one:
NARROW NO-BREAK SPACEi, decimal: &#8239; UTF-8 (e2, 80, af)

The last file is a page that has some references that had been
problematic in the past.

Change-Id: I92d3fedfdff33794ecd7be64ec4e2c84bf9073bd