Dev: make compare-extracts script accept a default page list
This list is a list of curated pages suitable for testing the summary
extracts. This is a cross-wiki list, so the domain needs to be set
for each entry.

In the script code mostly changed lang to domain variables and some
minor restructuring.

To run the script with the default list just omit the command line
parameter.

Change-Id: I9abc7a6fd9bf1ca92f8218e26e98cb0d27099bb2