Update mobile-sections page issues response structure
Our current rule of thumb is to return HTML in 'html' properties and plain
text content in 'text' properties.  The current page issues output in
mobile-sections-lead violates that rule by providing HTML content in a
'text' property.  This updates the response to provide the HTML in a new
'html' property, and the element's textContent in the 'text' property.

The text property is retained to aid with backward compatibility; any
clients relying on the existence of a 'text' property will still find it,
though its semantic meaning will have changed.

N.B.  We are not aware of any particular clients actually relying on this.

Change-Id: I164d5e5a26bc0cac6396b5b1da3a5e50d2770fd5