References: add section headings
Adds a section heading object if there is nothing else
in the section besides the heading and the reference list.

To keep the code simple (here and for clients) we just add direct parent
section headings, no grandparents.

Have to ignore white space when checking if the section is empty since
Parsoid pages often contain `\n`.

Updated diff test and spec to reflect the changes in references output.

Bug: T184327
Bug: T170690
Change-Id: I92bb263317e700a02f736f19d9d00ffe8126749d