Fix duplicate usage examples
On the English Wiktionary, make use of the microformats2
compatible HTML if present to parse the data (T138709).

This change adds `parsedExamples`
to the API response which contains objects composed of
(`example`, `translation`?, `literally`?).

To stay compatible with current clients the examples string
array is left unchanged and just contains the `example` part
of `parsedExamples`.


Bug: T187430
Change-Id: Ia62675bda62b0edf0e1789fd4117479c94fabea7