Add /mobile/app/bundle endpoint for legacy app bundles
Adds a bundle endpoint that responds with the equivalent of the bundles
currently used in the apps: a curated set of constant styles as well as
the Mobile.css for the site from which it's requested.  For consistency
with the existing bundles (see T190091), should be requested from
en.wikipedia.org.

Bug: T188919
Change-Id: If9c6fd8819c6f0d7040a5cd9d1064e66dc9fa941