References: make section_heading part of reference_list
Move section_heading to the corresponding reference_list object.
Only add reference_list if the section is empty after the heading and
reference_list content are removed.
We still want to keep all the references_by_id entries, regardless of
whether they show up in a reference list or not since we need that quick
lookup still to work.

Bug: T190219
Change-Id: I68e677e23e11c0bd3bd9908b545ce37b2e346aa5