Media: Decode file page titles before querying for imageinfo
Fixes this error: https://phabricator.wikimedia.org/P7016, which is caused
in turn by this rejection in the MW API:
https://phabricator.wikimedia.org/P7017

The underlying cause is that titles that we parse from HTML may contain
percent-encoded characters, but the MediaWiki API does not allow percent-
encoded characters in titles passed into the query module. So we'll need
to decode them before sending along to the API for our metadata query.

The error can be triggered in production by requesting
https://en.wikipedia.org/api/rest_v1/page/media/Yahoo!

Change-Id: I45eaa0bb1caedfbbff1a8acd1744ef3eca30f3ee