mobile-section + summary: prefer local description
If the local wiki sets a short description via the {{short description}}
magic word then use that instead of the description from Wikidata.
The preference of the local wiki is the default for the new
MediaWiki API prop=description[1], which this patch uses instead of the
pagterms API.

This patch also adds a new field `description_source` to indicate if the
description came from the `local` Wikipedia wiki or the `central`
Wikidata wiki.

Bumped the minor version for both endpoints to account for the
significant change of the values and the new `description_source` field.

[1] https://en.wikipedia.org/w/api.php?action=help&modules=query%2Bdescription

Bug: T191869
Change-Id: Ifb4546922313044e60a3988f627df928c7150f54