Add Android prod RL sync survey announcement
Adds an announcement to logged in production Android app users with
reading list sync enabled to take a survey about the new reading list
sync feature.

This survey is supposed to run for one week starting tomorrow morning.

Removed the payload response check for the announcement endpoint since
this leads to Icinga warnings when the survey has ended.

Bug: T191538
Change-Id: Ic06b20ec8dae2c7bb18e82c9ce6fe7531494907a