Media: Provide html and plain text image descriptions
Since this and other fields from Commons are freeform and may contain
HTML, and clients want to know exactly what they're getting, provide
both HTML and plain text versions.

Adds the 'striptags' npm package to achieve this, since this issue is
likely to keep coming up. Striptags is a package to strip HTML artifacts
that appears to be a mature project and does not use unsafe regexes.

Bumps the patch version to reflect the output change.

Change-Id: Icc4fe7040866d758b4c8f1264796741369670fb4