Summary: Don't strip space before period
Split up space stripping regex near punctuation by Latin and non-Latin
version. In the Latin version we now also require that more whitespace
or an HTML tag follows. In the non-Latin case this is not needed since
non-Latin period and comma already include space in the character
itself.

Bonus: corrected case of the word "Latin", and added an entry for
.NET Framework to the summary test page list file, which is used for
comparing extracts of different implementations.

Bumped the summary patch version.

Bug: T196523
Change-Id: Ic5dc87a2a7c3ee1b70ba0009d18ae2b1b7fcdec2