Media: Call RESTBase and MW API in parallel to improve performance
Makes the two major API calls in parallel rather than in series in an
effort to improve the performance of this endpoint.

This requires a couple of noteworthy changes:

* A mwApiQuery method is added to api-util.  This is because the MW API
  call for image info, which relies on `generator=images`, requires
  continuation for large pages (containing 50+ unique images). The new
  method handles query continuation.

* Calling the MW API initially removes our ability to specify a revision,
  so the spec is updated to no longer advertise support for specific
  revisions or tids.  This should not be a problem for clients in
  practice.  The minor version for the media endpoint is bumped to reflect
  the change.  The diff result fixture is also updated accordingly.

Note: When testing locally with ab I found that this performs better for
most pages, but slightly worse for pages requiring query continuation.
It's probably worth trying out to see the net effect in production (and
note that the penalty for query continuation may be less there, since
we'll be making internal API calls rather than hitting public endpoints).

Bug: T189830
Change-Id: Ia8ba690943647ba6be81c05f9a646bd47d33d850