content-html: add section edit button for lead section, too
Also add an edit button for the lead section. This is a bit different
from the other sections since it doesn't have a heading. We still create
the edit buttons, of course, but in the case of the lead section there
is no heading wrapper (since there's no heading). So, instead of adding
the edit button as a child of the heading wrapper add it to the
beginning of the section.

Keeping this patch separate from the base edit button patch since this
part may change more likely than the base case. Currently only the iOS
app creates an edit button inside the WebView content, and they don't
have description editing enabled yet. The HTML structure is taken from
[1].
OTOH, the Android app doesn't have a web button for the lead section.
Instead this and the surrounding elements (lead image, title,
description) are all native UI components.

Special care has to be taken to also float the edit button since design
wants the text to wrap around the button currently. The float value
depends on if the document is in RTL (float: left) or
LTR (float: right). I wanted to use `float: inline-end` instead but this
doesn't seem to be supported by Chrome yet (but works in FF).

[1] https://github.com/wikimedia/wikipedia-ios/blob/develop/www/js/sections.js#L189-L198

Bug: T170584
Change-Id: Iaa9934a7001d77964039bbb2378aac4b58fd09fc