content-html: theme preparation transform
Run the classifyElements function of ThemeTransform. That's the part
of ThemeTransform that can be run server side and is page specific but
not client/theme specific.

The actual switching of themes is accomplished by adding something like
class="pagelib_theme_dark" or similar to the root <html> tag.
This last step can be done easily on the client side; or for testing can
be done manually in DevTools.

The part that is not so clear yet is what to do about the requirement
of the pagelib CSS to have an element with a class of "content" as an
ancestor.

Bug: T170584
Change-Id: I648b25a9a247cff61949964af0da7b1b3505b710