Hygiene: separate reference_list definition in spec.yaml
That seems to be the proper way to do that in Swagger spec.
At least it makes swagger-to-graphql --swagger=references.yaml a bit
happier (in the RESTBase repo). See also the corresponding RB PR.

https://swagger.io/docs/specification/data-models/data-types/#array

Also mark the reference list id as nullable. Unfortunately,
`nullable: true` doesn't work yet and adds an extra error to
`npm run swagger`. But type: ['string', 'null'] does work.
Maybe the nullable field is only for Swagger 3.0?

https://swagger.io/docs/specification/data-models/data-types/#null

Also stop marking section_heading as required since it may be null.

Change-Id: I38d1892f471d396a9f5f2a713d115bfc174befa5