mobile-html spec: CollapseTables is not included for now
The CollapseTables transform is not included in the transforms for now.
So update the documentation accordingly. In the future we might
reconsider adding this transformation. It might involve some I18N setup
or some other way to get the three strings for the collapsed tables
(one for infoboxes, another for other kinds of tables, and a close
label).

Bug: T198667
Change-Id: Ie85464823e01a8992687f3230c6122d15fcb9a5b