mobile-html: set custom CSP header
I started using the CSP header provided by Parsoid but modified for
our purposes:
* only need https URIs since the apps should be able to handle https
  just fine
* added script-src since we also reference JS here
* added meta.wikimedia.org for common JS and CSS

To make this easier and to avoid having to skip the checkHeaders test
I changed the test to be a bit more lenient wrt CSP headers.
As long as it's set amd starts with default-src it's happy.

TODO: need to allow also certain data: URIs for the buttons from the
wikimedia-page-library, so that e.g. the edit button shows up.

Bug: T200464
Change-Id: Iafd38db72ec054fbc69db08cd7bb1d7bd945e009