Add multi-lang functionality announcements for Android app
Now we have announcements with multiple active wikis. So, changed the
activeWiki variable to an array.
The Chinese one has language variants. But we cannot use language
variant features yet since the apps don't know how that should be
specified. Instead we use country codes as an approximation.

There are also different announcements based on what version of the
Android app is installed:
>= 235: main announcement
<= 234: prompt to upgrade

I used https://en.wikipedia.org/wiki/ISO_3166-1#Current_codes to look up
country codes. This part is difficult to test.

The call to action for the main announcement is just '#languages', which
triggers the Android app to open the language settings. Pretty nifty.

There's a minor issue with the call to action for updating the app in
that it first launches the browser from the Android app before going to
the Play Store.[1] There doesn't seem to be a built-in way in the app
right now to send an Intend to launch the Play Store with its page.

Planned times:
Deploy time: Thursday, Aug  9th at 17:00–18:00 UTC
Start time:  Thursday, Aug  9th at 20:00 UTC
End time:    Thursday, Aug 16th at 20:00 UTC

This implies we have to purge the cache[2] in RB for the announcement
endpoints of the active wikis:
     'en.wikipedia.org',
     'de.wikipedia.org',
     'ru.wikipedia.org',
     'zh.wikipedia.org'

[1] https://phabricator.wikimedia.org/T191640#4489169
[2] https://www.mediawiki.org/wiki/Wikimedia_Apps/Team/RESTBase_services_for_apps/Deployment_process#Purge_Varnish_cache

Bug: T191640
Change-Id: Ied72a59035a99613ae9470fbedcd61ee7b2f3fb1