Leverage spec checks for checking response headers
Drops the headers.js test utility in favor of definining headers in the
Swagger spec x-amples.

In the event we need to test specifically for the absence of a header, I
think we can do that by adding a special field to the x-ample object and
handling it in spec.js.

Bug: T198134
Change-Id: I003dd411602661eab3abcaeffaaba1997a6761af