Hygiene: Remove some unnecessary or redundant integration tests
The project contains a large number of tests that make live network
requests, making the test suite rather slow to complete. Some of these
(particularly those with descriptions like "X should have expected
properties") are fully redundant with spec checks, so many of them are
removed here. Some others are removed because they are redundant with
other network tests, or just don't make sense in some way.

Note: I tried to be conservative here, and this probably doesn't capture
all of the tests that could be removed.  It's more of a first pass.

See also related task T151091 (which this doesn't actually resolve, so I
won't tag it).

Change-Id: I5a786113ebc2f55d4bb952c376fd603747fdf76e