mobile-html: rewrite thumb widths
Adds a transform for PCS that attempts to update image elements to use
a larger-size thumbnail, if available. The original image width is
obtained from the data-file-width attribute (which is no longer stripped
in earlier transformations); then we attempt to rewrite the thumbnail URL
for 640px, or failing that, 320px. Candidate widths for rewriting are
appropriate sizes selected from the thumbnail prerender buckets in
operations-mediawiki-config.

In the event that the thumb source URL is rewritten, other attributes
are also rewritten accordingly.

Bug: T201380
Change-Id: I50060aa89727413ac0aa662820a9a478e0e50952