Hygiene: add unit test for _rmMwIdAttributes
Parsoid emits element ids with 'mw' prefix and a base64 encoded counter
value[1].
The encoded counter can have '-', as we found recently, but also '_'.[2]
The latter was already covered by the regex since it's included in the
word character class.

This patch just adds a few unit test to make sure this behavior stays
the same when we make prod code changes.

[1] https://www.mediawiki
.org/wiki/Parsoid/MediaWiki_DOM_spec/Element_IDs#ID_prefixing_/_pattern
[2] https://phabricator.wikimedia
.org/diffusion/GPAR/browse/master/lib/utils/jsutils.js$166

Change-Id: I426f4c426662feadf70ddc02adab77fdf4cd41b1