Hygiene: remove random/summary
RB exposes it's own /page/random/summary based on /page/random/title
+ a hydrated /page/summary. So, this endpoint is not necessary anymore.
The only test case where this makes sense would be when testing a new
version of the randomizer with the apps but that is rare and then one
could also have a local RB instance as well.

Follow-up of I5a786113eb.

Another reason is that the new spec validation is probably too strict.
Not every (random) page has a thumbnail. This already let to one
failure:
https://integration.wikimedia.org/ci/job/mobileapps-deploy-npm-node-6-docker/60/console

Change-Id: I7a56e03b8aea65821011398a80e1c82270cc07d5