mobile-sections: move style tags in ref lists to end
References on enwiki are using template styles, which breaks parsing of
references in the Android app. There is a page library PR to fix that
but that wouldn't cover older installations.

This patch moves all <style> elements inside reference lists to the end
of the document, while avoid duplicate styles. The deduplication is not
exhaustive, though, since this would miss any styles outside of ref
lists already before this transformation but it should get us 95% there
(TM).

This is taking advantage of the data-mw-deduplicate attributes generated
by TemplateStyles. If there are other style elements present not having
these they will be simply removed.

Bumped mobile-sections patch version to 0.14.4.

Bug: T206527
Change-Id: I9a6a6c173141ec2c4379f5bd8e3f1fb2451b8616