mobile-html: add page header
Adds a <header> tag with the header, including title, possible button
for title pronunciation, possible description, and an edit button.
Most of the functionality is provided by the new method
EditTransform.newEditLeadSectionHeader in the wikimedia-page-library
PR 166, released as 6.2.0.

Removed the section edit button on the lead section since instead the
header provides it's own section edit button.

The call to add the page header requires metadata not present in the
Parsoid output. So, a parallel call to the MW API is made and the actual
addition of the header is delayed until both Parsoid and MW API
responses are obtained.

Bug: T201383
Change-Id: Ife46653c0ce0bf03ee414d41550f8617f117c912